--- a/32blit/graphics/mode7.hpp
+++ b/32blit/graphics/mode7.hpp
@@ -3,11 +3,11 @@
 #include <cstdint>
 
 #include "surface.hpp"
-#include "../types/map.hpp"
+#include "tilemap.hpp"
 
 namespace blit {
 
-  void mode7(Surface *dest, Surface *tiles, MapLayer *layer, float fov, float angle, Vec2 pos, float near, float far, Rect viewport);
+  void mode7(Surface *dest, TransformedTileLayer *layer, float fov, float angle, Vec2 pos, float near, float far, Rect viewport);
   Vec2 world_to_screen(Vec2 w, float fov, float angle, Vec2 pos, float near, float far, Rect viewport);
 
 }