… .github/workflows/Emscripten.yml
… .github/workflows/Visual Studio.yml
… .github/workflows/cmake.yml
… 32blit/CMakeLists.txt
… 32blit/graphics/tilemap.hpp
… 32blit/types/map.cpp
… 32blit/types/map.hpp
… vs/32blit/32blit.vcxproj