@Gadgetoid 2024-09-26
Niiice! I should drag my brain kicking and screaming back into the world of the productive and try to port Rocks & Diamonds.